internal/acctest: Return cluster member names when checking for clustering
Cluster member names are essential for testing TF provider on LXD cluster,
as we need to compare whether the resource was deployed on a correct node.

The clustering precheck already checks whether clustering is enabled and skips
the test otherwise. Now, it also retrieves the list of cluster member names
and skips the test if cluster member count is too low.

Signed-off-by: Din Music <din.music@canonical.com>